--- src/services/actions/actions.js
@@ -1,94 +1,0 @@
-import {
-  getDatafromApi,
-  getOrderDatafromApi,
-  burgerApiUrl,
-} from "../../utils/Api";
-
-export const GET_DATA_REQUEST = "GET_DATA_REQUEST";
-export const GET_DATA_SUCCESS = "GET_DATA_SUCCESS";
-export const GET_DATA_FAILED = "GET_DATA_FAILED";
-
-export const SET_ORDER_REQUEST = "SET_ORDER_REQUEST";
-export const SET_ORDER_SUCCESS = "SET_ORDER_SUCESS";
-export const SET_ORDER_FAILED = "SET_ORDER_FAILED";
-export const CLOSE_ORDER_MODAL = "CLOSE_ORDER_MODAL";
-
-export const SET_USER_DATA_REQUEST = "SET_USER_DATA_REQUEST";
-export const SET_USER_DATA_SUCCESS = "SET_USER_DATA_SUCCESS";
-export const SET_USER_DATA_FAILED = "SET_USER_DATA_FAILED";
-export const SET_USER_DATA_LOGOUT = "SET_USER_DATA_LOGOUT";
-export const SET_USER_DATA_EMAIL_TOKEN = "SET_USER_DATA_EMAIL_TOKEN";
-
-export const OPEN_MODAL_INGRIDIENT = "OPEN_MODAL_INGRIDIENT";
-export const CLOSE_MODAL_INGRIDIENT = "CLOSE_MODAL_INGRIDIENT";
-
-export const ADD_BUN_TO_CONSTRUCTOR = "ADD_BUN_TO_CONSTRUCTOR";
-export const ADD_INGRIDIENT_TO_CONSTRUCTOR = "ADD_INGRIDIENT_TO_CONSTRUCTOR";
-export const REMOVE_INGRIDIENT_FROM_CONSTRUCTOR =
-  "REMOVE_INGRIDIENT_FROM_CONSTRUCTOR";
-export const MOVE_INGRIDIENT = "START_MOVE_INGRIDIENT";
-export const CLEAR_BIN_CONSTRUCTOR = "CLEAR_BIN_CONSTRUCTOR";
-
-export function getData() {
-  return function (dispatch) {
-    dispatch({
-      type: GET_DATA_REQUEST,
-    });
-    getDatafromApi(burgerApiUrl)
-      .then((res) => {
-        if (res && res.success) {
-          dispatch({
-            type: GET_DATA_SUCCESS,
-            res: res
-          });
-        } else {
-          dispatch({
-            type: GET_DATA_FAILED,
-          });
-        }
-      })
-      .catch((err) => {
-        console.log(`ошибка ${err}`);
-      });
-  };
-}
-
-export function setOrder(idArr) {
-  return function (dispatch) {
-    dispatch({
-      type: SET_ORDER_REQUEST,
-    });
-    getOrderDatafromApi(burgerApiUrl, idArr)
-      .then((res) => {
-        if (res && res.success) {
-          dispatch({
-            type: SET_ORDER_SUCCESS,
-            res: res,
-          });
-          dispatch({ type: CLEAR_BIN_CONSTRUCTOR });
-        } else {
-          dispatch({
-            type: SET_ORDER_FAILED,
-          });
-        }
-      })
-      .catch((err) => {
-        console.log(`ошибка ${err}`);
-      });
-  };
-}
-
-export function addBunToConstructor(bun) {
-  return {
-    type: ADD_BUN_TO_CONSTRUCTOR,
-    payload: bun,
-  };
-}
-
-export function addIngridientToConstructor(ingridient, uuid) {
-  return {
-    type: ADD_INGRIDIENT_TO_CONSTRUCTOR,
-    payload: ingridient,
-    uuid: uuid,
-  };
-}
