+++ src/services/reducers/auth.js
@@ -1,0 +1,63 @@
+import {
+  SET_USER_DATA_REQUEST,
+  SET_USER_DATA_SUCCESS,
+  SET_USER_DATA_FAILED,
+  SET_USER_DATA_LOGOUT,
+  SET_USER_DATA_EMAIL_TOKEN,
+} from "../actions/auth";
+
+const initUserData = {
+  request: false,
+  failed: false,
+  isLoggin: false,
+  emailToken: false,
+  user: {
+    email: "",
+    name: "",
+  },
+};
+
+export const userData = (state = initUserData, action) => {
+  switch (action.type) {
+    case SET_USER_DATA_REQUEST: {
+      return { ...state, request: true };
+    }
+    case SET_USER_DATA_SUCCESS: {
+      return {
+        ...state,
+        request: false,
+        failed: false,
+        isLoggin: action.res.success,
+        user: action.res.user,
+      };
+    }
+    case SET_USER_DATA_FAILED: {
+      return {
+        ...state,
+        request: false,
+        failed: true,
+      };
+    }
+    case SET_USER_DATA_EMAIL_TOKEN: {
+      return {
+        ...state,
+        emailToken: true,
+      };
+    }
+
+    case SET_USER_DATA_LOGOUT: {
+      return {
+        ...state,
+        request: false,
+        isLoggin: !action.res.success,
+        user: {
+          name: "",
+          email: "",
+        },
+      };
+    }
+
+    default:
+      return state;
+  }
+};
