+++ src/services/reducers/reducers.js
@@ -1,223 +1,12 @@
 import { combineReducers } from "redux";
 import update from "immutability-helper";
-
-import {
-  GET_DATA_REQUEST,
-  GET_DATA_SUCCESS,
-  GET_DATA_FAILED,
-  SET_ORDER_REQUEST,
-  SET_ORDER_SUCCESS,
-  SET_ORDER_FAILED,
-  SET_USER_DATA_REQUEST,
-  SET_USER_DATA_SUCCESS,
-  SET_USER_DATA_FAILED,
-  CLOSE_ORDER_MODAL,
-  OPEN_MODAL_INGRIDIENT,
-  CLOSE_MODAL_INGRIDIENT,
-  ADD_BUN_TO_CONSTRUCTOR,
-  ADD_INGRIDIENT_TO_CONSTRUCTOR,
-  REMOVE_INGRIDIENT_FROM_CONSTRUCTOR,
-  MOVE_INGRIDIENT,
-  CLEAR_BIN_CONSTRUCTOR,
-  SET_USER_DATA_LOGOUT,
-  SET_USER_DATA_EMAIL_TOKEN,
-} from "../actions/actions";
-
-const initStateData = {
-  data: [],
-  dataRequest: false,
-  dataFailed: false,
-  success: false,
-};
-
-const initOrderData = {
-  openModal: false,
-  name: "",
-  order: {
-    number: "",
-  },
-  success: false,
-  orderRequest: false,
-  orderFailed: false,
-};
-
-const initIngridientModal = {
-  openModal: false,
-  data: {},
-};
-
-const initConstructor = {
-  bun: null,
-  ingridients: [],
-};
-
-const initUserData = {
-  request: false,
-  failed: false,
-  isLoggin: false,
-  emailToken: false,
-  user: {
-    email: "",
-    name: "",
-  },
-};
-
-export const getDataReducer = (state = initStateData, action) => {
-  switch (action.type) {
-    case GET_DATA_REQUEST: {
-      return { ...state, dataRequest: true };
-    }
-    case GET_DATA_SUCCESS: {
-      return {
-        ...state,
-        data: action.res.data,
-        dataRequest: false,
-        success: action.res.success,
-      };
-    }
-    case GET_DATA_FAILED: {
-      return { ...state, dataFailed: true, dataRequest: false };
-    }
-    default:
-      return state;
-  }
-};
-
-export const submitOrderData = (state = initOrderData, action) => {
-  switch (action.type) {
-    case SET_ORDER_REQUEST: {
-      return { ...state, orderRequest: true };
-    }
-    case SET_ORDER_SUCCESS: {
-      return {
-        ...state,
-        order: { ...state.order, number: action.res.order.number },
-        success: action.res.success,
-        orderRequest: false,
-        openModal: true,
-      };
-    }
-    case SET_ORDER_FAILED: {
-      return { ...state, orderFailed: true, orderRequest: false };
-    }
-    case CLOSE_ORDER_MODAL: {
-      return { state: initOrderData };
-    }
-
-    default:
-      return state;
-  }
-};
-
-export const ingridientData = (state = initIngridientModal, action) => {
-  switch (action.type) {
-    case OPEN_MODAL_INGRIDIENT: {
-      return {
-        openModal: true,
-        data: action.payload,
-      };
-    }
-    case CLOSE_MODAL_INGRIDIENT: {
-      return { state: initIngridientModal };
-    }
-    default:
-      return state;
-  }
-};
-
-export const constructorBin = (state = initConstructor, action) => {
-  switch (action.type) {
-    case ADD_BUN_TO_CONSTRUCTOR: {
-      return {
-        ...state,
-        bun: action.payload,
-      };
-    }
-    case ADD_INGRIDIENT_TO_CONSTRUCTOR: {
-      return {
-        ...state,
-        ingridients: [
-          ...state.ingridients,
-          { ...action.payload, uuid: action.uuid },
-        ],
-      };
-    }
-    case REMOVE_INGRIDIENT_FROM_CONSTRUCTOR: {
-      return {
-        ...state,
-        ingridients: state.ingridients.filter(
-          (elem) => elem.uuid !== action.uuid
-        ),
-      };
-    }
-    case MOVE_INGRIDIENT: {
-      return update(state, {
-        ingridients: {
-          $splice: [
-            [action.drag, 1],
-            [action.hover, 0, state.ingridients[action.drag]],
-          ],
-        },
-      });
-    }
-    case CLEAR_BIN_CONSTRUCTOR: {
-      return initConstructor;
-    }
-
-    default:
-      return state;
-  }
-};
-
-export const userData = (state = initUserData, action) => {
-  switch (action.type) {
-    case SET_USER_DATA_REQUEST: {
-      return { ...state, request: true };
-    }
-    case SET_USER_DATA_SUCCESS: {
-      return {
-        ...state,
-        request: false,
-        failed: false,
-        isLoggin: action.res.success,
-        user: action.res.user,
-      };
-    }
-    case SET_USER_DATA_FAILED: {
-      return {
-        ...state,
-        request: false,
-        failed: true,
-      };
-    }
-    case SET_USER_DATA_EMAIL_TOKEN: {
-      return {
-        ...state,
-        emailToken: true,
-      };
-    }
-
-    case SET_USER_DATA_LOGOUT: {
-      return {
-        ...state,
-        request: false,
-        isLoggin: !action.res.success,
-        user: {
-          name: "",
-          email: "",
-        },
-      };
-    }
-
-    default:
-      return state;
-  }
-};
+import { getDataReducer } from "./ingridients";
+import { submitOrderData, constructorBin } from "./order";
+import { userData } from "./auth";
 
 export const rootReducer = combineReducers({
   data: getDataReducer,
   order: submitOrderData,
-  ingridient: ingridientData,
   constructorBin: constructorBin,
   userData: userData,
 });
