+++ src/utils/Api.js
@@ -2,19 +2,23 @@ import { getCookie } from "./cookie";
 
 export const burgerApiUrl = "https://norma.nomoreparties.space/api/";
 
-const checkResponse =(res) => {
+const checkResponse = (res) => {
   if (res.ok) {
     return res.json();
   }
   return Promise.reject(`Ошибка: ${res.status}`);
-}
+};
+
+const request = (url, options) => {
+  return fetch(url, options).then(checkResponse);
+};
 
 export function getDatafromApi(burgerApiUrl) {
-  return fetch(`${burgerApiUrl}ingredients`).then(checkResponse);
+  return request(`${burgerApiUrl}ingredients`);
 }
 
 export function getOrderDatafromApi(burgerApiUrl, idArr) {
-  return fetch(`${burgerApiUrl}orders`, {
+  return request(`${burgerApiUrl}orders`, {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
@@ -22,116 +26,105 @@ export function getOrderDatafromApi(burgerApiUrl, idArr) {
     body: JSON.stringify({
       ingredients: idArr,
     }),
-  })
-  .then(checkResponse);
+  });
 }
 
-
-export const sendRegistrationUsertoApi = async (burgerApiUrl, data) => {
-  return await fetch(`${burgerApiUrl}auth/register`, {
+export const sendRegistrationUsertoApi = (burgerApiUrl, data) => {
+  return request(`${burgerApiUrl}auth/register`, {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
     },
     body: JSON.stringify({
-      "email": data.email, 
-      "password": data.password, 
-      "name": data.name 
-  } )
-  })
-  .then(checkResponse)
-}
+      email: data.email,
+      password: data.password,
+      name: data.name,
+    }),
+  });
+};
 
-export const setloginUserApi = async (burgerApiUrl, data) => {
-  return await fetch(`${burgerApiUrl}auth/login`, {
+export const setloginUserApi = (burgerApiUrl, data) => {
+  return request(`${burgerApiUrl}auth/login`, {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
     },
     body: JSON.stringify({
-      "email": data.email,
-      "password": data.password
-    })
-  })
-  .then(checkResponse)
-}
+      email: data.email,
+      password: data.password,
+    }),
+  });
+};
 
-export const passwordForgotApi = async (burgerApiUrl, data) => {
-  return await fetch(`${burgerApiUrl}password-reset`, {
+export const passwordForgotApi = (burgerApiUrl, data) => {
+  return request(`${burgerApiUrl}password-reset`, {
     method: "POST",
     headers: {
-      "Content-Type": "application/json"
+      "Content-Type": "application/json",
     },
     body: JSON.stringify({
-      "email" : data
-    })
-  })
-  .then(checkResponse)
-}
+      email: data,
+    }),
+  });
+};
 
-export const passwordResetApi = async (burgerApiUrl, data) => {
-  return await fetch(`${burgerApiUrl}password-reset/reset`, {
+export const passwordResetApi = (burgerApiUrl, data) => {
+  return request(`${burgerApiUrl}password-reset/reset`, {
     method: "POST",
     headers: {
-      "Content-Type": "application/json"
+      "Content-Type": "application/json",
     },
     body: JSON.stringify({
-      "password" : data.password,
-      "token": data.token
-    })
-  })
-  .then(checkResponse)
-}
+      password: data.password,
+      token: data.token,
+    }),
+  });
+};
 
-export const getUserApi = async (burgerApiUrl) => {
-  return await fetch(`${burgerApiUrl}auth/user`, {
+export const getUserApi = (burgerApiUrl) => {
+  return request(`${burgerApiUrl}auth/user`, {
     method: "GET",
     headers: {
       "Content-Type": "application/json",
-      Authorization: "Bearer " + getCookie('accessToken')
-    }
-  })
-  .then(checkResponse)
-}
+      Authorization: "Bearer " + getCookie("accessToken"),
+    },
+  });
+};
 
-export const changeUserDataApi = async (burgerApiUrl, data) => {
-  return await fetch(`${burgerApiUrl}auth/user`, {
+export const changeUserDataApi = (burgerApiUrl, data) => {
+  return request(`${burgerApiUrl}auth/user`, {
     method: "PATCH",
     headers: {
       "Content-Type": "application/json",
-      Authorization: "Bearer " + getCookie('accessToken')
+      Authorization: "Bearer " + getCookie("accessToken"),
     },
     body: JSON.stringify({
-      "name": data.name,
-      "email": data.email
-    })
-  })
-  .then(checkResponse)
-}
+      name: data.name,
+      email: data.email,
+    }),
+  });
+};
 
-
-export const refreshTokenApi = async (burgerApiUrl) => {
-  return await fetch(`${burgerApiUrl}auth/token`, {
+export const refreshTokenApi = (burgerApiUrl) => {
+  return request(`${burgerApiUrl}auth/token`, {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
     },
-    body: JSON.stringify ({
-      "token": getCookie('refreshToken')
-    })
-  })
-  .then(checkResponse)
-}
+    body: JSON.stringify({
+      token: getCookie("refreshToken"),
+    }),
+  });
+};
 
-export const logoutUserApi = async (burgerApiUrl) => {
-  return await fetch(`${burgerApiUrl}auth/logout`, {
+export const logoutUserApi = (burgerApiUrl) => {
+  return request(`${burgerApiUrl}auth/logout`, {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
     },
-    body: JSON.stringify ({
-      "token": getCookie('refreshToken')
-    })
-  })
-  .then(checkResponse)
-}
+    body: JSON.stringify({
+      token: getCookie("refreshToken"),
+    }),
+  });
+};
